gpu/requirements.txt
redis>=4.5.0
rq==1.13.0
requests
pillow>=9.0.0
